Move SSO config to separate file
